+++ sdk/communication/Azure.Communication.Administration/src/PhoneNumberAdministrationClient.cs
@@ -21,11 +21,15 @@ namespace Azure.Communication.Administration
         internal ClientDiagnostics ClientDiagnostics { get; private set; }
         internal PhoneNumberAdministrationRestClient RestClient { get; }
 
-        /// <summary>
-        /// Initializes a phone number administration client with an Azure resource connection string.
-        /// </summary>
-        public PhoneNumberAdministrationClient(string connectionString)
-            : this(new PhoneNumberAdministrationClientOptions(), ConnectionString.Parse(connectionString))
+        /// <summary> Initializes a new instance of <see cref="PhoneNumberAdministrationClient"/>.</summary>
+        /// <param name="endpoint">The URI of the Azure Communication Services resource.</param>
+        /// <param name="keyCredential">The <see cref="AzureKeyCredential"/> used to authenticate requests.</param>
+        /// <param name="options">Client option exposing <see cref="ClientOptions.Diagnostics"/>, <see cref="ClientOptions.Retry"/>, <see cref="ClientOptions.Transport"/>, etc.</param>
+        public PhoneNumberAdministrationClient(Uri endpoint, AzureKeyCredential keyCredential, CommunicationIdentityClientOptions? options = default)
+            : this(
+                AssertNotNull(endpoint, nameof(endpoint)),
+                options ?? new CommunicationIdentityClientOptions(),
+                AssertNotNull(keyCredential, nameof(keyCredential)))
         { }
 
         /// <summary>
@@ -34,7 +38,7 @@ namespace Azure.Communication.Administration
         public PhoneNumberAdministrationClient(string connectionString, PhoneNumberAdministrationClientOptions? options = default)
             : this(
                   options ?? new PhoneNumberAdministrationClientOptions(),
-                  ConnectionString.Parse(connectionString))
+                  ConnectionString.Parse(AssertNotNull(connectionString, nameof(connectionString))))
         { }
 
         internal PhoneNumberAdministrationClient(PhoneNumberAdministrationClientOptions options, ConnectionString connectionString)
@@ -45,6 +49,15 @@ namespace Azure.Communication.Administration
         {
             RestClient = new PhoneNumberAdministrationRestClient(clientDiagnostics, pipeline, endpointUrl);
             ClientDiagnostics = clientDiagnostics;
+        }
+
+        internal PhoneNumberAdministrationClient(Uri endpoint, CommunicationIdentityClientOptions options, AzureKeyCredential credential)
+        {
+            ClientDiagnostics = new ClientDiagnostics(options);
+            RestClient = new PhoneNumberAdministrationRestClient(
+                ClientDiagnostics,
+                options.BuildHttpPipeline(credential),
+                endpoint.AbsoluteUri);
         }
 
         /// <summary>Initializes a new instance of <see cref="PhoneNumberAdministrationClient"/> for mocking.</summary>
@@ -943,6 +956,13 @@ namespace Azure.Communication.Administration
                 scope.Failed(ex);
                 throw;
             }
+        }
+
+        private static T AssertNotNull<T>(T argument, string argumentName)
+            where T : class
+        {
+            Argument.AssertNotNull(argument, argumentName);
+            return argument;
         }
     }
 }