+++ sdk/communication/Azure.Communication.Administration/tests/Infrastructure/CommunicationIdentityClientTestEnvironment.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using Azure.Core.TestFramework;
 
 namespace Azure.Communication.Administration.Tests
@@ -8,11 +9,12 @@ namespace Azure.Communication.Administration.Tests
     public class CommunicationIdentityClientTestEnvironment: TestEnvironment
     {
         internal const string ConnectionStringEnvironmentVariableName = "COMMUNICATION_CONNECTION_STRING";
-
         public string ConnectionString => GetRecordedVariable(ConnectionStringEnvironmentVariableName);
 
-        internal const string EndpointStringEnvironmentVariableName = "COMMUNICATION_ENDPOINT_STRING";
+        internal const string EndpointEnvironmentVariableName = "COMMUNICATION_ENDPOINT";
+        public Uri Endpoint => new Uri(GetRecordedVariable(EndpointEnvironmentVariableName));
 
-        public string EndpointString => GetRecordedVariable(EndpointStringEnvironmentVariableName);
+        internal const string AccessKeyEnvironmentVariableName = "COMMUNICATION_ACCESS_KEY";
+        public string AccessKey => GetRecordedVariable(AccessKeyEnvironmentVariableName);
     }
 }