+++ sdk/communication/Azure.Communication.Sms/tests/SmsClientTestEnvironment.cs
@@ -7,10 +7,10 @@ namespace Azure.Communication.Sms.Tests
 {
     public class SmsClientTestEnvironment : TestEnvironment
     {
-        internal const string ConnectionStringEnvironmentVariableName = "AZURE_COMMUNICATION_LIVETEST_CONNECTION_STRING";
-        internal const string PhoneNumberEnvironmentVariableName = "AZURE_PHONE_NUMBER";
-
+        internal const string ConnectionStringEnvironmentVariableName = "COMMUNICATION_CONNECTION_STRING";
         public string ConnectionString => GetRecordedVariable(ConnectionStringEnvironmentVariableName);
+
+        internal const string PhoneNumberEnvironmentVariableName = "AZURE_PHONE_NUMBER";
         public string PhoneNumber => GetRecordedVariable(PhoneNumberEnvironmentVariableName);
     }
 }