+++ sdk/communication/Azure.Communication.Administration/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - Added support to create CommunicationIdentityClient with TokenCredential
+- Added support to create CommunicationIdentityClient with AzureKeyCredential
+- Added support to create PhoneNumberAdministrationClient with AzureKeyCredential
 ### Fixed
 - Issue with paging results not pulling next pages